test(bridge,tests): rename/split residual real-I/O tests onto the seam boundary
- bridge/pending-compact: asserts the no-token-on-disk invariant against real
  LocalStorage disk bytes (its own header says "real LocalStorage … actual
  disk bytes, not an in-memory hash") — rename to *.integration.test.js.
- tests/env-setup: spawns real `bun scripts/env-setup.js` and asserts the
  written .env plus chmod 0600 file mode — rename to *.integration.test.js.
- tests/capture-cli-golden: split the pure applyTransforms unit test from the
  capture/verify block, which spawns a real stub bin and compares golden files
  on disk → capture-cli-golden.integration.test.js.

Leaves the SC3 allow-list as guard self-tests that scan/spawn the real repo by
design: services/oauth/no-github (rg over source), tests/check-ambient-deps,
tests/check-subprocess-in-tests. librc/manager-* keep their injected
execSync/spawn fakes (no real subprocess; the check-subprocess guard passes).

Refs spec 0640 § B (Part 02 sweep; closes the SC3 rg -l mkdtemp residual to zero).
