+++ tests/capture-cli-golden.test.js
@@ -1,14 +1,7 @@
-import { test, describe, beforeEach, afterEach } from "node:test";
+import { test, describe } from "node:test";
 import assert from "node:assert/strict";
-import { mkdtempSync, writeFileSync, rmSync } from "node:fs";
-import { join } from "node:path";
-import { tmpdir } from "node:os";
 
-import {
-  applyTransforms,
-  capture,
-  verify,
-} from "../scripts/capture-cli-golden.mjs";
+import { applyTransforms } from "../scripts/capture-cli-golden.mjs";
 
 describe("applyTransforms", () => {
   test("normalises matches with the g flag", () => {
@@ -16,57 +9,5 @@ describe("applyTransforms", () => {
       { pattern: "RUN[A-Z]+", replacement: "STAMP" },
     ]);
     assert.equal(out, "ts=STAMP ts=STAMP");
-  });
-});
-
-describe("capture + verify against a stub bin", () => {
-  let dir;
-  let stub;
-
-  beforeEach(() => {
-    dir = mkdtempSync(join(tmpdir(), "golden-"));
-    stub = join(dir, "stub.js");
-    // Deterministic except a timestamp token the transform normalises.
-    writeFileSync(
-      stub,
-      `process.stdout.write("hello stamp=" + "RUN-ABC123\\n");`,
-    );
-    writeFileSync(
-      join(dir, "cases.json"),
-      JSON.stringify([
-        {
-          name: "hello",
-          args: [],
-          exitCode: 0,
-          stdoutFile: "hello.stdout.txt",
-          stderrFile: "hello.stderr.txt",
-          transform: [
-            {
-              pattern: "stamp=[A-Z0-9-]+",
-              replacement: "stamp=NORMALIZED",
-            },
-          ],
-        },
-      ]),
-    );
-  });
-
-  afterEach(() => rmSync(dir, { recursive: true, force: true }));
-
-  const flags = () => ({ bin: "stub", exec: stub, goldenDir: dir });
-
-  test("capture then verify is clean", () => {
-    const captured = capture(flags());
-    assert.equal(captured.count, 1);
-    const result = verify(flags());
-    assert.equal(result.ok, true, JSON.stringify(result.diffs));
-  });
-
-  test("verify detects a drift when the bin output changes", () => {
-    capture(flags());
-    writeFileSync(stub, `process.stdout.write("changed\\n");`);
-    const result = verify(flags());
-    assert.equal(result.ok, false);
-    assert.ok(result.diffs.length > 0);
   });
 });
